entitlements

+++ lib/bencher_json/src/organization/usage.rs
@@ -4,11 +4,9 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::BigInt;
-
 #[typeshare::typeshare]
 #[derive(Debug, Clone, Deserialize, Serialize)]
 #[cfg_attr(feature = "schema", derive(JsonSchema))]
 pub struct JsonUsage {
-    pub metrics_used: BigInt,
+    pub metrics_used: u32,
 }
